fixed some issues in fee types, mentoring invoice

See merge request walayatkhan/nbeac!179

+++ database/migrations/2020_09_15_112646_create_mentoring_invoices_table.php
@@ -25,7 +25,7 @@ class CreateMentoringInvoicesTable extends Migration
                 ->references('id')
                 ->on('departments');
             $table->string('slip')->nullable();
-            $table->integer('fee_type_id')->unsigned();
+            $table->bigInteger('fee_type_id')->unsigned();
             $table->foreign('fee_type_id')
                 ->references('id')
                 ->on('fee_types');
@@ -67,5 +67,6 @@ class CreateMentoringInvoicesTable extends Migration
     public function down()
     {
         Schema::dropIfExists('mentoring_invoices');
+
     }
 }
